Engine rework part 5 - Rework camera - Re add mouse/keyboard movements
- Create class camera controller to decouple unrelated logic from camera class
- Update gameState to use this class, fixing onUpdate gameloop logic related with camera movement and renderer class logic related with projection view matrix
- Re add camera movement
- Update shader vert to accept projectionview together, rename model to transform
- Remove base.h class
- Add key to enable mouse cursor while pushed
- Enable minimize in gameloop
- Move RenderApi call to onKeyPressed
- Simplify event classes
- Move not engine keyboard events to Sandbox
- Add recursive searching for source files
- Disable imgui code for now

+++ potatocraft/src/core/input.h
@@ -14,7 +14,7 @@ namespace potatoengine {
 
 		static bool IsMouseButtonPressed(MouseCode button);
 		static glm::vec2 GetMousePosition();
-		static float GetMouseX();
-		static float GetMouseY();
+		static double GetMouseX();
+		static double GetMouseY();
 	};
 }
